pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webapp-example-backend"
version = "0.1.0"
description = "A web application backend example"
requires-python = ">=3.11"
authors = [
    { name = "Joseph McGuire", email = "joseph.c.mcg@gmail.com" }
]
license = { text = "BSD 3-Clause" }

dependencies = [
    "Flask==3.1.0",
    "requests==2.32.3",
    "unittest2==1.1.0",
    "numpy==2.0.2",
    "pandas==2.2.3",
    "scikit-learn==1.5.2",
    "matplotlib==3.9.2",
    "seaborn==0.13.2",
    "setuptools==65.5.0",
    "tqdm==4.67.0",
    "shap==0.46.0",
    "pyyaml==6.0.2",
    "flask_cors==5.0.0",
    "blinker==1.9.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "cloudpickle==3.1.0",
    "colorama==0.4.6",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "fonttools==4.55.0",
    "pyparsing==3.2.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.2",
    "scipy==1.14.1",
    "six==1.16.0",
    "slicer==0.0.8",
    "threadpoolctl==3.5.0",
    "traceback2==1.4.0",
    "tzdata==2024.2",
    "Werkzeug==3.1.3",
    "dotenv",
]

[tool.setuptools]
package-dir = {"" = "app"}

[tool.setuptools.packages.find]
where = ["app"]

[project.scripts]
serve_model = "run.py:main"